Cherry picked part of commit 824adca to port over *only* the JackSupport refactoring.
