+++ client/jackson/src/main/kotlin/net/corda/client/jackson/internal/JacksonUtils.kt
@@ -1,18 +1,13 @@
 package net.corda.client.jackson.internal
 
+import com.fasterxml.jackson.annotation.JacksonAnnotationsInside
 import com.fasterxml.jackson.core.JsonGenerator
 import com.fasterxml.jackson.core.JsonParser
-import com.fasterxml.jackson.databind.JsonDeserializer
 import com.fasterxml.jackson.databind.JsonNode
-import com.fasterxml.jackson.databind.JsonSerializer
 import com.fasterxml.jackson.databind.ObjectMapper
-import com.fasterxml.jackson.databind.module.SimpleModule
+import com.fasterxml.jackson.databind.annotation.JsonSerialize
+import com.fasterxml.jackson.databind.ser.std.ToStringSerializer
 import com.fasterxml.jackson.module.kotlin.convertValue
-
-inline fun <reified T : Any> SimpleModule.addSerAndDeser(serializer: JsonSerializer<in T>, deserializer: JsonDeserializer<T>) {
-    addSerializer(T::class.java, serializer)
-    addDeserializer(T::class.java, deserializer)
-}
 
 inline fun JsonGenerator.jsonObject(fieldName: String? = null, gen: JsonGenerator.() -> Unit) {
     fieldName?.let { writeFieldName(it) }
@@ -24,3 +19,7 @@ inline fun JsonGenerator.jsonObject(fieldName: String? = null, gen: JsonGenerato
 inline fun <reified T> JsonParser.readValueAs(): T = readValueAs(T::class.java)
 
 inline fun <reified T : Any> JsonNode.valueAs(mapper: ObjectMapper): T = mapper.convertValue(this)
+
+@JacksonAnnotationsInside
+@JsonSerialize(using = ToStringSerializer::class)
+annotation class ToStringSerialize
